watchtower-plugin/requirements.txt
pyln-client
requests
coincurve
cryptography
pyzbase32
plyvel